working 3D version

+++ src/Core/functions_CPU/RingWeights_core.h
@@ -43,7 +43,7 @@ float RingWeights_main(float *residual, float *weights, int horiz_window_halfsiz
 /************2D functions ***********/
 float RingWeights2D(float *residual, float *weights, int horiz_window_halfsize, int vert_window_halfsize, int full_window, long anglesDim, long detectorsDim, long i, long j);
 /************3D functions ***********/
-float RingWeights3D(float *residual, float *weights, int horiz_window_halfsize, int vert_window_halfsize, int slices_window_halfsize, int full_window, long anglesDim, long detectorsDim, long slices, long i, long j, long k);
+float RingWeights3D(float *residual, float *weights, long horiz_window_halfsize, long vert_window_halfsize, long slices_window_halfsize, int full_window, long anglesDim, long detectorsDim, long slices, long i, long j, long k);
 #ifdef __cplusplus
 }
 #endif
